[refactor] rename and edit files
