+++ src/container/main/router.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createBrowserRouter } from 'react-router-dom';
-import SingUpPage from '../user/signup';
-import EditUserPage from '../user/edit';
+import SingUpPage from '../user/SignUp';
+import EditUser from '../user/EditUser';
 import ErrorPage from '../ErrorPage';
 import { ErrorPageCode } from '../ErrorPage/types';
 import { getNavigationsValue } from '@brojs/cli';
@@ -20,7 +20,7 @@ export const router = createBrowserRouter([
       },
       {
         path: getNavigationsValue('sberhubproject.editUser'),
-        element: <EditUserPage />
+        element: <EditUser />
       },
       {
         path: getNavigationsValue('sberhubproject.eventList'),
