+++ src/container/user/UserForm/Photo/types.ts
@@ -1,5 +1,5 @@
 export interface PhotoProps {
   id: string;
   name: string;
-  defaultPhoto: string | null;
+  defaultPhoto?: string;
 }
