--- src/container/user/form/button/index.style.ts
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-import Button from '@mui/material/Button';
-
-export const ButtonStyled = styled(Button)`
-  background-color: var(--tg-theme-button-color);
-  color: var(--tg-theme-button-text-color);
-`;
