--- src/container/user/form/button/types.ts
@@ -1,6 +1,0 @@
-export interface ButtonProps {
-  children: React.ReactNode;
-  id: string | null;
-  name: string | null;
-  className?: string | null;
-}
